Replace flake8 & isort with ruff linter / upgrade mypy & its config and move to pyproject.toml
