+++ .pre-commit-config.yaml
@@ -1,18 +1,13 @@
 repos:
   - repo: https://github.com/ambv/black
-    rev: 22.8.0
+    rev: 23.1.0
     hooks:
     - id: black
-  - repo: https://github.com/pycqa/isort
-    rev: 5.11.5
+  - repo: https://github.com/charliermarsh/ruff-pre-commit
+    rev: v0.0.241
     hooks:
-    - id: isort
-      args: ["--profile", "black"]
-  - repo: https://github.com/pycqa/flake8
-    rev: 5.0.4
-    hooks:
-    - id: flake8
+    - id: ruff
   - repo: https://github.com/pre-commit/mirrors-mypy
-    rev: v0.982
+    rev: v0.991
     hooks:
       - id: mypy
