+++ setup.cfg
@@ -34,49 +34,3 @@ classifiers =
 zip_safe = False
 packages = find_namespace:
 python_requires = >=3.7
-test_suite = tests
-
-[aliases]
-test=pytest
-
-[flake8]
-statistics = true
-ignore = E203, E501, W503, B950
-select = B,C,E,F,W,T4,B9
-
-
-[mypy]
-show_error_codes = True
-ignore_missing_imports = True
-
-# Disallow dynamic typing
-#disallow_any_decorated = True
-#disallow_any_unimported = True
-#disallow_any_expr = True
-#disallow_any_explicit = True
-disallow_any_generics = True
-#disallow_subclassing_any = True
-
-# Untyped definitions and calls
-disallow_untyped_calls = True
-disallow_untyped_defs = True
-disallow_incomplete_defs = True
-disallow_untyped_decorators = True
-
-# None and Optional handling
-no_implicit_optional = True
-strict_optional = True
-
-# Configuring warnings
-warn_redundant_casts = True
-warn_unused_ignores = True
-warn_no_return = True
-warn_return_any = True
-warn_unreachable = True
-
-# Miscellaneous strictness flags
-implicit_reexport = True
-strict_equality = True
-
-[mypy-tests.*]
-ignore_errors = True
